Final Commit
I renamed some of the files. Added a video player along with controls and captions.  Created a text file and filled it with the captions for the video along with time stamps of where each caption should go. Added styles to affect the spacing of the video player and the colors of the captions. Finally, I added comments for my code. I should mention that the first four seconds of the video did not function at all. Ready for review.